requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# MongoDB Integration
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
cryptography==41.0.7

# AI & ML
google-generativeai==0.3.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2

# NLP & Text Processing
nltk==3.8.1
textblob==0.17.1

# Data Visualization & Analytics
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0
locust==2.20.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Monitoring & Logging
sentry-sdk==1.38.0
psutil==5.9.6

# Task Queue (for async AI processing)
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
hiredis==2.2.3

# Performance & Caching
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Rate Limiting
django-ratelimit==4.1.0